renaming the files and reference for better readability
